[feature/member-login] 회원 로그인 구현

+++ member-api/member-application/src/main/java/org/flab/deliveryplatform/member/application/service/WithdrawMemberService.java
@@ -1,6 +1,7 @@
 package org.flab.deliveryplatform.member.application.service;
 
 import lombok.RequiredArgsConstructor;
+import org.flab.deliveryplatform.member.application.port.EncryptManager;
 import org.flab.deliveryplatform.member.application.port.MemberRepository;
 import org.flab.deliveryplatform.member.application.port.WithdrawMemberUseCase;
 import org.flab.deliveryplatform.member.application.port.dto.WithdrawMemberCommand;
@@ -14,11 +15,16 @@ public class WithdrawMemberService implements WithdrawMemberUseCase {
 
     private final MemberRepository memberRepository;
 
+    private final EncryptManager encryptManager;
+
     @Override
     public void withdraw(WithdrawMemberCommand withdrawMemberCommand) {
-        Member member = memberRepository.findByEmailAndPassword(
-                withdrawMemberCommand.getEmail(), withdrawMemberCommand.getPassword())
+        Member member = memberRepository.findByEmail(withdrawMemberCommand.getEmail())
             .orElseThrow(InvalidMemberInfoException::new);
+
+        if (!encryptManager.isMatch(withdrawMemberCommand.getPassword(), member.getPassword())) {
+            throw new InvalidMemberInfoException();
+        }
 
         memberRepository.delete(member);
     }
